requirements.txt
torch==2.2.2
torchvision==0.17.2
Pillow>=10.0.0
numpy==1.26.4
faiss-cpu>=1.7.4
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9
tqdm>=4.66.0
transformers>=4.40.0,<5.0.0
huggingface-hub>=0.20.0
scipy>=1.11.0
opencv-python-headless>=4.8.0
